refactor: update drawertitle
